@@ -10,18 +10,17 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^17.0.8",
-    "@angular/common": "^17.0.8",
-    "@angular/compiler": "^17.0.8",
-    "@angular/core": "^17.0.8",
-    "@angular/forms": "^17.0.8",
-    "@angular/localize": "^17.0.8",
-    "@angular/platform-browser": "^17.0.8",
-    "@angular/platform-browser-dynamic": "^17.0.8",
-    "@angular/router": "^17.0.8",
+    "@angular/animations": "^18.0.1",
+    "@angular/common": "^18.0.1",
+    "@angular/compiler": "^18.0.1",
+    "@angular/core": "^18.0.1",
+    "@angular/forms": "^18.0.1",
+    "@angular/localize": "^18.0.1",
+    "@angular/platform-browser": "^18.0.1",
+    "@angular/platform-browser-dynamic": "^18.0.1",
+    "@angular/router": "^18.0.1",
     "@jest/globals": "^28.0.0",
     "chart.js": "^4.4.1",
-    "hammerjs": "^2.0.0",
     "ng2-charts": "^5.0.4",
     "primeng": "^17.3.2",
     "rxjs": "^7.8.1",
@@ -29,16 +28,16 @@
     "zone.js": "~0.14.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.0.8",
-    "@angular/cli": "^17.0.8",
-    "@angular/compiler-cli": "^17.0.8",
+    "@angular-devkit/build-angular": "^18.0.2",
+    "@angular/cli": "^18.0.2",
+    "@angular/compiler-cli": "^18.0.1",
     "@types/jasmine": "~4.3.0",
     "@types/jest": "^29.2.4",
     "autoprefixer": "^10.4.13",
     "jest": "^29.7.0",
-    "jest-preset-angular": "^13.1.4",
+    "jest-preset-angular": "^14.1.0",
     "postcss": "^8.4.21",
     "ts-jest": "^29.1.1",
-    "typescript": "~5.2.2"
+    "typescript": "~5.4.5"
   }
 }
